requirements_TSC.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.4.0
beartype==0.22.5
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fastai==2.8.5
fastcore==1.8.16
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
idna==3.11
imbalanced-learn==0.14.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plum-dispatch==2.6.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorchtime==0.1.2
pyts==0.13.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
smart_open==7.5.0
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.8
threadpoolctl==3.6.0
torch==2.4.1
torchaudio==2.9.0
torchmetrics==1.8.2
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
tsai==0.4.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.2
wrapt==2.0.1
yarl==1.22.0
